MonkeyPatch: only write when feature is new.
Always writing is causing errors when you're using Rollout in a case
where you want certain parts to only read from slaves and other parts to
only write to your master.

As we're overwriting the standard `Rollout.active?` class, checking for
the `active` state in a read-only slave would cause a Redis error. Only
writing this feature when it doesn't exist resolves this issue.